restructured and method for roles added
