+++ backend/flower/core/README.md
@@ -14,22 +14,20 @@
 
 ### <a name="Create"></a> Создание
 
-Если создаёте часть ядра для новой сущности, создайте отдельный пакет, 
-в пакет могут входить файлы `models.py`, `resources.py`, `utils` и
-обязательно входит файл `__init__.py`. 
+Если создаёте часть ядра для новой сущности, создайте отдельную папку, 
+в папку могут входить файлы `models.py`, `resources.py`, `utils`. 
 
-- После создания модели, она сначала
-добавляется в `<название пакета>/__init__.py`, а затем в `core/models.py`.
+- После создания модели, она добавляется `core/models.py`.
 Импортировать модели также стоит из `core.models.py`. 
 
 - После создания ресурса нужо добавить в конец файла переменную `routes`, 
-заполните её роутами. Затем следует добавить ресурс в `__init__`.
+заполните её роутами. Затем следует добавить ресурс в `core/resources.py`.
 
 Пример:
 
 Нужно сделать роуты для пользователей.
 
-1. Создаём пакет `users`.
+1. Создаём папку `users`.
 2. Создаём модель `UserModel`:
 
 ```python
@@ -44,25 +42,16 @@ class UserModel(db.Model):
     password = db.Column(db.String(120), nullable=False)
 ```
 
-3. Добавляем модель в `users/__init__.py`:
+3. Добавляем модель в `core/models.py`:
 
 ```python
-from .models import UserModel
+from .users.models import UserModel
 
 __all__ = ['UserModel']
 
 ```
 
-4. Добавляем модель в `core/models.py`:
-
-```python
-from .users import UserModel
-
-__all__ = ['UserModel']
-
-```
-
-5. Создаём эндпойт или функцию:
+4. Создаём эндпойт или функцию:
 
 ```python
 from starlette.endpoints import HTTPEndpoint
@@ -84,37 +73,27 @@ async def ping(request):
 ```
 
 
-6. Добавляем его в переменную `routes`:
+5. Добавляем его в переменную `routes`:
 
 ```python
 routes = [
     Route('/', Users),
-    Route('/', ping, methods=['GET'])
+    Route('/ping', ping, methods=['GET'])
 ]
 ```
 
 обратите внимание на путь эндпойнта `'/'`. Префикс эндпойнта не нужно указывать
 при роутинге в вашем ресурсе, он будет указан в дальнейшем.
 
-7. Добавляем роуты к моделе в `users/__init__.py`:
-
-```python
-from .models import UserModel
-from .resources import routes
-
-__all__ = ['routes', 'UserModel']
-
-```
-
-8. Добавляем роуты юзера к общим роутам:
+6. Добавляем роуты юзера к общим роутам:
 
 ```python
 from starlette.routing import Mount
 
-from . import users
+from .users.resources import routes as users_routes
 
 routes = [
-    Mount('/users', routes=users.routes),
+    Mount('/users', routes=users_routes),
 ]
 
 __all__ = ['routes']
