+++ backend/flower/core/__init__.py
@@ -1,15 +1,4 @@
-from starlette.routing import Route, Mount
-from starlette.responses import JSONResponse
+from .database import db
+from .routes import routes
 
-from . import users
-
-
-async def ping(request):
-    return JSONResponse({'onPing': 'wePong'})
-
-routes = [
-    Route('/ping', ping),
-    Mount('/users', routes=users.routes),
-]
-
-__all__ = ['routes']
+__all__ = ['db', 'routes']
