+++ backend/flower/core/database.py
@@ -1,5 +1,5 @@
 from gino_starlette import Gino
-from flower.config import DB_HOST, DB_USER, DB_DATABASE, DB_PASSWORD, DB_PORT
+from ..config import DB_HOST, DB_USER, DB_DATABASE, DB_PASSWORD, DB_PORT
 
 
 db = Gino(
