+++ backend/flower/core/models.py
@@ -1,5 +1,5 @@
-from .users import UserModel
-from .roles import RoleModel
-from .permissions import PermissionModel
+from .roles.models import RoleModel
+from .users.models import UserModel
+from .permissions.models import PermissionModel
 
 __all__ = ['UserModel', 'RoleModel', 'PermissionModel']
