--- backend/flower/core/permissions/__init__.py
@@ -1,3 +1,0 @@
-from .models import PermissionModel
-
-__all__ = ['PermissionModel']
