--- backend/flower/core/roles/__init__.py
@@ -1,3 +1,0 @@
-from .models import RoleModel
-
-__all__ = ['RoleModel']
