--- backend/flower/core/users/__init__.py
@@ -1,4 +1,0 @@
-from .models import UserModel
-from .resources import routes
-
-__all__ = ['routes', 'UserModel']
